+++ src/main/java/model/game/dto/GameDto.java
@@ -1,4 +1,4 @@
-package model.game.dao;
+package model.game.dto;
 
 public record GameDto(long gameId, String turn) {
 }
